requirements.txt
aiocontextvars==0.2.2
aiofiles==0.4.0
aiosqlite==0.10.0
alembic==1.0.11
aniso8601>=3.0,<=6.0.1
astroid==2.2.5
authy==2.2.3
asyncpg
bcrypt==3.1.7
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
contextvars==2.4
databases==0.7.0
dataclasses==0.6
decorator==4.4.0
dnspython==1.16.0
fastapi==0.95.2
httptools==0.1.1
h11==0.8.1
idna==2.8
immutables==0.9
isort==4.3.21
itsdangerous==1.1.0
lazy-object-proxy==1.4.1
Mako==1.0.13
MarkupSafe>=2.0
mccabe==0.6.1
nbconvert==5.4.1
nejma==0.1.0
passlib==1.7.1
pbr==5.4.0
promise==2.2.1
psycopg2-binary
pycparser==2.19
pydantic==1.10.12
PyJWT==1.7.1
pylint==2.3.1
PySocks==1.7.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
Rx==1.6.1
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.4.53
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
starlette==0.27.0
Tempita==0.5.2
twilio==6.29.0
ujson==1.35
urllib3==1.25.3
uvicorn>=0.23.0
websockets>=11.0
wincertstore==0.2
wrapt==1.11.2
gunicorn
fastapi-mail==1.3.0
fastapi_pagination
python-decouple
jinja2
python-multipart
aiosmtplib
email-validator==1.3.1
google-auth
google-auth-oauthlib
google-auth-httplib2
firebase-admin
